% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pythagorean.R
\name{Pythagorean}
\alias{Pythagorean}
\title{A Pythagorean Theorem Function}
\usage{
Pythagorean(a, b)
}
\arguments{
\item{a}{A length of one side.}

\item{b}{A length of another side.}
}
\description{
This function allows you to obtain the hypoteneuse of a triangle.
}
\examples{
Pythagorean()

}
\keyword{Pythagorean}
