% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trimmed_Mean.R
\name{trimmedMean}
\alias{trimmedMean}
\title{A Trimmed Mean Function}
\usage{
trimmedMean(x, s = 0, l = 0)
}
\arguments{
\item{x}{A vector.}

\item{s}{How many of the lowest numbers are to be trimmed away.}

\item{l}{How many of the largest numbers are to be trimmed away.}
}
\description{
This function allows you to obtain a trimmed mean of a vector.
}
\examples{
trimmedmean()

}
\keyword{trimmed}
